moved token fetching to provider
